Week 11/requirements.txt
mcp
mcp[cli]
python_dotenv
agents
requests
sqllite
flask
werkzeug